Make chaining interceptors a type safe operation
An interceptor may depend on others. In such cases, the dependencies must
come before it in the interceptor's chain, but the framework wasn't
enforcing such a dependency, making chaining interceptors error prone
and putting the burden of checking for possible errors on the user.

With these changes, an interceptor's constructor now enforce its
dependencies at the type level, leaving the compiler to guarantee it.

To do so, the interceptor chain was converted to a linked list in
reverse order (from newest to oldest, like in functional languages). The
constructor of a interceptor is such that it specifies the interface
it expects for the interceptors coming before it in the chain.